Remove data augmentation
Removed data augmentation as it changed the classweights
